Update README.md
This course is about inverse problems in imaging. The mathematical reconstruction and processing of images is of fundamental importance in state of the art applications in health and geosciences, e.g. in medical tomography, in high-resolution microscopy or in geophysical inversion. In many cases, underlying inverse problems can be formulated and solved using variational methods and partial differential equations. This course offers a theoretical as well as an applied insight into inverse problems and variational methods for mathematical imaging. It addresses reconstruction problems of different imaging modalities (e.g. CT or PET) in biomedicine and geophysics. The course covers the full chain of solving inverse problems in imaging, namely